[cfe] Copy default values into redirecting factories
Redirecting factories are not allowed to declare default values but
should instead inherit the default values from their target. Since
the current encoding of redirecting factories is a callable function
used as the tear-off by the VM, the default values should be copied
to the procedure representing the redirecting factory.

Closes https://github.com/dart-lang/sdk/issues/48548

Change-Id: I0e601cd4920bc1059c9bf99e4a4c9c0e20c183e6
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/237160
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>