[infra] Add MSVC windows to test matrix
Includes precompiled build targets to ensure we try to build most of the
C++ code.

Does not include any tests. We assume that if the build succeeds the
executables will behave the same. If this turns out to not be the case,
we could add a tests later.

Bug: https://github.com/dart-lang/sdk/issues/48544

Change-Id: I5f6b74676d9bb705d29303c8597d4f420f9030d5
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/237340
Reviewed-by: Alexander Thomas <athom@google.com>
Commit-Queue: Daco Harkes <dacoharkes@google.com>