[vm] Avoid wasted space in the data array for map and set constants.
The element data array in the VM's hash map/set implementation always
has a power-of-two length, growing exponentially as entries/elements are
added.

For constant maps/sets, this is not necessary, and the unused parts of
these data arrays is just wasted space.

This changes the data arrays for constant maps and sets to only contain
the actual entries/elements of the map/set.

Tested: ci
Change-Id: I061eb44b6c74c87157ef920143bf1885582f1b9a
Cq-Include-Trybots: luci.dart.try:vm-kernel-precomp-linux-debug-x64-try
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/237140
Reviewed-by: Daco Harkes <dacoharkes@google.com>