Rework handling of `identical` in the `InvocationInferrer` class hierarchy.
Rather than have a separate subclass of `MethodInvocationInferrer` to
handle `identical`, with a method override for `_iterateArguments`
that makes calls to flow analysis in between handling arguments, we
now have an `_isIdentical` method to determine whether the flow
analysis calls are needed, and the logic for calling into flow
analysis is directly in the base class implementation of
`resolveInvocation`.

This change will be necessary to support
https://github.com/dart-lang/language/issues/731 (improved type
inference for functions like `Iterable.fold`) because it will require
changing the order in which we visit arguments sometimes, so the
technique of overriding `_iterateArguments` won't work anymore.

Bug: https://github.com/dart-lang/language/issues/731
Change-Id: I77bd20b10494da941a1bc8b5fc63968c6c2c68cc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/238120
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>