[CFE] Make macros work when initializing from dill
Previously, when initializing from dill, if the dill we initialize from
has a macro in it (and we don't remove it because of invalidation) the
macro system would think it was a precompiled dill and thus not compile
it as something it could use for running macros. Then when it tried
to use it, it would figure out it didn't have it (and crash).

This CL fixes the issue by reworking the logic:
* Assume the platform is there (dart: uris)
* Assume the uris given specifically as precompiled (and not removed
  because of invalidation in/by the incremental compiler) is there.

No tests are added, but it has been manually tested in flutter.

Change-Id: Ica6c4f2f0601334d50beccdd7cf9d5740e611384
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/239026
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Jens Johansen <jensj@google.com>