[dart2js] Fix nested conditions in global inference
A bug in global inference made the analysis accidentally carry over
the state of a nested condition in a subexpression and conclude
that it was part of the state of the outer condition.

This was due to the fact that we didn't clear the state when
popping out of the context and reentering the conition state.

Fixes https://github.com/dart-lang/sdk/issues/48571

Change-Id: I4e8294bb1a05bb3b910b6aec374357d75acb67cb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/240321
Reviewed-by: Stephen Adams <sra@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>