[cfe] Run incremental macro tests in separate incremental compilers
The way constants are printed using ast-to-text depend on which
libraries are available in the component. Because the incremental
compiler (by default) only returns the libraries that have been
recompiled, the first compilation contains more libraries than the
subsequent compilations. Using the same incremental compiler throughout
the incremental macro test made the output depend on the order in which
the tests were run, which in turn was dependent on the file order in
the directory.

To avoid this, and to support running individual tests with a stable
output, each test is now using its own increment compiler for all the
incremental steps within the test.

Closes https://github.com/dart-lang/sdk/issues/48558

Change-Id: I37946ed0f940b6644c5efb2b7d167b2a5717f2f3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/240721
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>