Rework exception handling for legacy handlers
This is an attempt to make the change to exception handling that you
mentioned in a previous CL. Let me know if this doesn't resolve the
problem you saw.

Also, this changes the number of responses and notifications found in a
poorly written test. I don't know whether it will change the behavior
that clients see, and if so whether that will cause problems, but it's
a little concerning. Interested to hear your thoughts.

Change-Id: I135e96bea5bd6e5e673dcdddd1b421e4c2011f89
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/240647
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>