[cfe] Ensure default values are cloned in order of dependency
Default values were cloned in order of where the need for the cloning
was seen. This meant that declaration order could introduce cases where
the default values of a dependent constructor wasn't created before it
was cloned.

This CL registers the cloners with the member that they operate on and
uses this to ensure that default values are cloned in order of
dependency.

Closes https://github.com/dart-lang/sdk/issues/48769

Change-Id: Ie4fad010ba8273d9287911336d30cf47fc3e031e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/240844
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>