Require '@dart = 2.15' for binary_protocol.dart
Otherwise it fails internally, because the package is not null safe
in google3, and cannot be made because it has dependencies
(compiler and dev_compiler) that are not null safe. But the binary
protocol only talks to null safe code, so can be upgraded to the
newer language version.

Bug: https://buganizer.corp.google.com/issues/229124206
Change-Id: I215799de7d40a673c4c11d4d3a9e1c936b979df0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/241143
Reviewed-by: Jake Macdonald <jakemac@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>