[dart2js] Avoid some 'allowNull' options
This change is in preparation for migrating dart2js to NNBD.

In serialization some methods have an 'allowNull' or 'emptyAsNull'
optional arguments.

I'm trying to remove these named parameters and replace

    // @dart = 2.10
    Foo readFoo({bool allowNull: false});

with

    // TODO: @dart = 2.16
    Foo readFoo();
    Foo? readFooOrNull();

Having two functions with different return types helps keep the
nullability 'tight'. This helps the migration tool infer the correct
types.

For collections, we don't need two functions on the 'write' side, but
I think it is useful to keep the readers and writers consistent.

Change-Id: I40dc216b50c90e9bacf487436605f0bf2b1d94e8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/241206
Reviewed-by: Nate Biggs <natebiggs@google.com>
Reviewed-by: Joshua Litt <joshualitt@google.com>
Commit-Queue: Stephen Adams <sra@google.com>