[dart2js] Remove most of elements/operators.dart
The rich hierarchy of operators is no longer relevant since parsing
was moved to the common front-end.

All that remains is a Set of operator names that can be used to
declare instance methods.

This set is used for picking the 'operator' variety of Selector. I'm not
sure we even need operator selectors, but I am leaving that for future
consideration.

Change-Id: I7337c0cf82a3cb8f7ebf5fa17a737bac3a83f48b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/241560
Reviewed-by: Joshua Litt <joshualitt@google.com>
Commit-Queue: Stephen Adams <sra@google.com>