[vm] Fix assertion in ICData::AssertInvariantsAreSatisfied()
The ICData::AssertInvariantsAreSatisfied() method will grab a snapshot
of the ICData's entrie()s backing store array and does verifications
on it.

Though the place when we check for the last entry (which is a back-ref)
we re-load the entries() array which could be updated at this point.

This causes us to possibly fail an assertion which checks that for empty
ic data arrays it should be a cached empty array.

Closes https://github.com/dart-lang/sdk/issues/48787

TEST=Should fix flaky assertion hit.

Change-Id: I45cf57bd13a7c2e49c395eb25a2f6789c5140bf9
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/242109
Reviewed-by: Alexander Aprelev <aam@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>